+++ src/umbraco.cms/businesslogic/member/Member.cs
@@ -109,7 +109,7 @@ namespace umbraco.cms.businesslogic.member
 
         public static IEnumerable<Member> GetAllAsList()
         {
-            int totalRecs;
+            long totalRecs;
             return ApplicationContext.Current.Services.MemberService.GetAll(0, int.MaxValue, out totalRecs)
                 .Select(x => new Member(x))
                 .ToArray();
@@ -154,7 +154,7 @@ namespace umbraco.cms.businesslogic.member
         /// <returns></returns>
         public static Member[] getMemberFromFirstLetter(char letter)
         {
-            int totalRecs;
+            long totalRecs;
 
             return ApplicationContext.Current.Services.MemberService.FindMembersByDisplayName(
                 letter.ToString(CultureInfo.InvariantCulture), 0, int.MaxValue, out totalRecs, StringPropertyMatchType.StartsWith)
@@ -164,7 +164,7 @@ namespace umbraco.cms.businesslogic.member
 
         public static Member[] GetMemberByName(string usernameToMatch, bool matchByNameInsteadOfLogin)
         {
-            int totalRecs;
+            long totalRecs;
             if (matchByNameInsteadOfLogin)
             {
                 var found = ApplicationContext.Current.Services.MemberService.FindMembersByDisplayName(
@@ -289,7 +289,7 @@ namespace umbraco.cms.businesslogic.member
             if (string.IsNullOrEmpty(email))
                 return null;
 
-            int totalRecs;
+            long totalRecs;
             var found = ApplicationContext.Current.Services.MemberService.FindByEmail(
                 email, 0, int.MaxValue, out totalRecs, StringPropertyMatchType.Exact);
